+++ formula/src/main/java/com/instacart/formula/Evaluation.kt
@@ -3,8 +3,8 @@ package com.instacart.formula
 /**
  * The result of [evaluate][Formula.evaluate] function.
  *
- * @param Output Usually a data class returned by formula that contains data and callbacks.
- * When it is used to render UI, we call it a render model (Ex: ItemRenderModel).
+ * @param Output Usually a data class returned by formula that contains data and event
+ * listeners. When it is used to render UI, we call it a render model (Ex: ItemRenderModel).
  *
  * @param updates A list of asynchronous events that formula wants to listen and respond to.
  */